+++ Runtime/Font/Font.cpp
@@ -74,8 +74,8 @@ namespace Directus
 
 		// Load font
 		vector<unsigned char> atlasBuffer;
-		int texAtlasWidth = 0;
-		int texAtlasHeight = 0;
+		unsigned int texAtlasWidth = 0;
+		unsigned int texAtlasHeight = 0;
 		if (!m_context->GetSubsystem<ResourceManager>()->GetFontImporter()._Get()->LoadFont(filePath, m_fontSize, atlasBuffer, texAtlasWidth, texAtlasHeight, m_glyphs))
 		{
 			LOG_ERROR("Font: Failed to load font \"" + filePath + "\"");
@@ -147,8 +147,8 @@ namespace Directus
 				int spaceOffset = m_glyphs[ASCII_SPACE].horizontalOffset;
 				int spaceCount = 8; // spaces in a typical terminal
 				int tabSpacing = spaceOffset * spaceCount;
-				int columnHeader = pen.x - position.x; // -position.x because it has to be zero based so we can do the mod below
-				float nextTabStop = tabSpacing - (columnHeader % tabSpacing);
+				int columnHeader = (int)pen.x - (int)position.x; // -position.x because it has to be zero based so we can do the mod below
+				int nextTabStop = tabSpacing - (columnHeader % tabSpacing);
 				pen.x += nextTabStop;
 				continue;
 			}
